Define `urlize' to handle file names on Windows that contain back slashes: they are converted to forward slashes to avoid problems in different names between the Emacs and Java sides. Invoke urlize on all the places dealing with filenames.
Similarly, use a dir separator, which is backslash on Windows and
forward slash on Unix.

Changed the message printed when typing is faster than the response
time of the Emacs/Java process pair. The message is more informative
now.

Use yes-or-no-p instead of the XEmacs specific
yes-or-no-p-maybe-dialog-box.

Use the ^ and $ begin and end delimiters.